refactor: 타입 및 상수를 constants로 이동 및 정리
