+++ src/basic/components/CartPage.tsx
@@ -1,5 +1,5 @@
-import type { CartItem, Coupon, NotificationVariant } from '../../types';
-import type { ProductWithUI } from '../constants';
+import type { CartItem, Coupon } from '../../types';
+import type { NotificationVariant, ProductWithUI } from '../constants';
 import { Icon } from './icons';
 import { calculateCartTotal } from '../models/cart';
 import { CartHeader } from './ui/CartHeader';
