+++ src/basic/constants/notification.ts
@@ -1,0 +1,7 @@
+export type NotificationVariant = 'error' | 'success' | 'warning';
+
+export interface Notification {
+  id: string;
+  message: string;
+  variant: NotificationVariant;
+}
