+++ src/basic/hooks/useCouponForm.ts
@@ -1,14 +1,8 @@
 import { useState } from 'react';
 
 import type { Coupon } from '../../types';
+import { initialCouponForm } from '../constants';
 import { useForm } from '../utils/hooks/useForm';
-
-const initialCouponForm: Coupon = {
-  name: '',
-  code: '',
-  discountType: 'amount',
-  discountValue: 0,
-};
 
 interface UseCouponFormProps {
   onAddCoupon: (coupon: Coupon) => void;
